+++ src/renderer.js
@@ -4,14 +4,18 @@ const Bus = require("./bus")
 const Autodect = require("./autodetect")
 const Session = require("./session")
 const Macros = require("./macros")
-//const CustomCSS = require("./storage/custom-css")
 const Theme = require("./storage/theme")
 const Settings = require("./settings")
 
 window.messages = window.messages || []
 
-Bus.on(Bus.events.CHANGE_THEME, (data) => {
-  Theme.changeTheme(data)
+const theme = Settings.get("theme")
+
+theme && Theme.changeTheme({ theme }).then(() => {})
+
+Bus.on(Bus.events.CHANGE_THEME, async (data) => {
+  await Theme.changeTheme(data)
+  document.body.classList.remove("loading")
 })
 
 m.mount(document.getElementById("sessions"), UI.Sessions)
@@ -47,11 +51,6 @@ Bus.on(Bus.events.FOCUS, (session) => {
   document.querySelector("title").innerText = session.name
   session.attach(document.getElementById("feed-wrapper"))
   m.redraw()
-
-  // Set theme from settings
-  // TODO: flashes original theme
-  const theme = Settings.get("theme")
-  Theme.changeTheme({ theme: theme })
 })
 
 Bus.on("macro", (macro) => {
